@@ -1,5 +1,6 @@
-class Index::CommentReply < ActiveRecord::Base
-  after_destroy :delete_thumb_up
+class Index::CommentReply < ApplicationRecord
+  after_destroy :delete_thumb_up, :clear_cache
+  store_accessor :info, :tbp_counts, :read_times # 点赞数/阅读次数
 
   belongs_to :user,
              class_name: 'Index::User',
@@ -24,6 +25,25 @@ class Index::CommentReply < ActiveRecord::Base
   validates :comment_id, presence: true
   validates :content, presence: true, length: { minimum: 1, maximum: 255 }
 
+  #----------------------------域------------------------------
+  default_scope { order('index_comment_replies.id DESC') }
+
+  def create cmt, user
+    ApplicationRecord.transaction do
+      self.comment = cmt
+      self.user = user
+      self.save!
+      cmt.update! rep_counts: (cmt.rep_counts || 0) + 1
+    end
+  end
+
+  def drop cmt
+    ApplicationRecord.transaction do
+      self.destroy!
+      cmt.update! rep_counts: (cmt.rep_counts || 0) - 1
+    end
+  end
+
   # ------------------------文件类型------------------------- #
   def file_type
     :replies
@@ -44,7 +64,16 @@ class Index::CommentReply < ActiveRecord::Base
     Index::ThumbUp.has?(self, user)
   end
 
+  # -----------------------点赞信息------------------------ #
+  def thumb_up_info(user)
+    Index::ThumbUp.counts_and_has?(self, user)
+  end
+
   def delete_thumb_up
     # Index::ThumbUp.destroy self
+  end
+
+  def clear_cache
+    Cache.new["reply_#{self.id}"] = nil
   end
 end
