@@ -1,4 +1,7 @@
-class Index::Workspace::EditComment < ActiveRecord::Base
+class Index::Workspace::EditComment < ApplicationRecord
+  after_update :update_cache
+  after_destroy :clear_cache
+
   belongs_to :user,
              class_name: 'Index::User',
              foreign_key: :user_id
@@ -11,4 +14,12 @@ class Index::Workspace::EditComment < ActiveRecord::Base
   validates :resource_type, presence: true, inclusion: { in: ['Index::Workspace::Article'] }
   validates :hash_key, presence: true, length: { maximum: 32 }
   validates :content, presence: true, length: { minimum: 1, maximum: 255 }
+
+  def update_cache
+    Cache.new["edit_comment_#{self.id}"] = self
+  end
+
+  def clear_cache
+    Cache.new["edit_comment_#{self.id}"] = nil
+  end
 end
