@@ -1,5 +1,5 @@
 class Index::CommentsController < IndexController
-  before_action :check_login, only: [:create, :destroy]
+  before_action :require_login, only: [:create, :destroy]
   before_action :set_resource
   before_action :set_comment, only: [:show, :destroy]
 
@@ -9,25 +9,21 @@ class Index::CommentsController < IndexController
     count = params[:count] || 15
     count = 100 if count.to_i > 100  # 限制返回的评论条数
     @nonpaged_comments = @resource ? @resource.comments : Index::Comment.none
-    @comments = @nonpaged_comments.page(page).per(count)
+    @comments = @nonpaged_comments.page(page).per(count).includes(:limit_3_replies)
   end
 
   # GET /index/comments/1
   # GET /index/comments/1.json
   def show
-    @replies = @comment ? @comment.replies.limit(15) : Index::Comment.none
+    # @replies = @comment ? @comment.replies.limit(15) : Index::Comment.none
   end
 
   # POST /:resource_type/:resource_id/comments
   def create
     @comment = Index::Comment.new(comment_params)
-
     unless @code
-      @comment.resource = @resource
-      @comment.user = @user
-      @code = @comment.save ? 'Success' : 'Fail'
+      @code = @comment.create(@resource, @user) ? 'Success' : 'Fail'
     end
-    @code ||= 'Fail'
     respond_to do |format|
       format.json { render :show, status: @comment.id.nil? ? :unprocessable_entity : :created }
     end
@@ -38,10 +34,11 @@ class Index::CommentsController < IndexController
   def destroy
     unless @code
       if @user = @comment.user || @user.can_edit?(:comment, @resource)  # 判断用户是否有权限删除评论
-        @code = @comment.destroy ? 'Success' : 'Fail'
+        @code = @comment.drop(@resource) ? 'Success' : 'Fail'
+      else
+        @code = 'NoPermission'
       end
     end
-    @code ||= 'Fail'
     render json: { code: @code }
   end
 
@@ -49,7 +46,7 @@ class Index::CommentsController < IndexController
 
   # Use callbacks to share common setup or constraints between actions.
   def set_comment
-    @comment = @resource.comments.find_by_id params[:id] if @resource
+    comment_cache params[:id] if @resource
     @code ||= 'ResourceNotExist' unless @comment
   end
 
@@ -63,9 +60,9 @@ class Index::CommentsController < IndexController
       resource_id = params[:resource_id]
       case resource_type
       when 'articles'
-        @resource = Index::Workspace::Article.shown.find_by_id resource_id
+        @resource = shown_article_cache resource_id
       when 'corpuses'
-        @resource = Index::Workspace::Corpus.shown.find_by_id resource_id
+        @resource = shown_corpus_cache resource_id
       end
       @code = 'ResourceNotExist' unless @resource
   end
