@@ -1,42 +1,37 @@
 class Index::Workspace::EditCommentsController < IndexController
-  before_action :check_login
+  before_action :require_login
   before_action :set_resource
+  before_action :check_permission, only: [:add_reply, :remove_reply, :destroy]
   before_action :set_edit_comment, only: [:show, :add_reply, :remove_reply, :destroy]
 
   def index
     @edit_comments = @resource ? @resource.edit_comments : Index::Workspace::EditComment.none
   end
 
-  def show
-  end
+  def show; end
 
   def create
     @edit_comment = Index::Workspace::EditComment.new(edit_comment_params)
 
-    unless @code
-      @edit_comment.user = @user
-      @edit_comment.resource = @resource
-      @code = 'Success'if @edit_comment.save
-    end
+    @edit_comment.user = @user
+    @edit_comment.resource = @resource
+    @code = @edit_comment.save ? 'Success' : 'Fail'
 
-    @code ||= 'Fail'
     respond_to do |format|
       format.json { render :show, status: @edit_comment.id.nil? ? :unprocessable_entity : :created }
     end
   end
 
   def add_reply
-    unless @code
-      if params[:reply] && params[:reply].size < 255
-        reply = {
-          user_id: @user.id,
-          content: params[:reply]
-        }
-        @edit_comment.replies[Time.now.strftime('%Y-%m-%d %H:%M:%S')] = reply
-        @code = @edit_comment.save ? 'Success' : 'Fail'
-      else
-        @edit_comment.errors[:base] = 'size of reply should be 1-255'
-      end
+    if params[:reply] && params[:reply].size < 255
+      reply = {
+        user_id: @user.id,
+        content: params[:reply]
+      }
+      @edit_comment.replies[Time.now.strftime('%Y-%m-%d %H:%M:%S')] = reply
+      @code = @edit_comment.save ? 'Success' : 'Fail'
+    else
+      @edit_comment.errors[:base] = 'size of reply should be 1-255'
     end
     @code ||= 'Fail'
     respond_to do |format|
@@ -45,19 +40,17 @@ class Index::Workspace::EditCommentsController < IndexController
   end
 
   def remove_reply
-    unless @code
-      key = params[:hash_key]
-      if @edit_comment.replies.key?(key)
-        # 允许删除回复的条件
-        if @edit_comment.replies[key]['user_id'] == @user.id || # 1: 回复者
-           @edit_comment.user == @user || # 2: 评论者
-           @user.can_edit?(:delete_comment, @resource) # 3. 具有删除评论权限的用户
+    key = params[:hash_key]
+    if @edit_comment.replies.key?(key)
+      # 允许删除回复的条件
+      if @edit_comment.replies[key]['user_id'] == @user.id || # 1: 回复者
+         @edit_comment.user == @user || # 2: 评论者
+         @user.can_edit?(:delete_comment, @resource) # 3. 具有删除评论权限的用户
 
-          @edit_comment.replies.delete key
-          @code = @edit_comment.save ? 'Success' : 'Fail'
-        else
-          @code = 'NoPermission'
-        end
+        @edit_comment.replies.delete key
+        @code = @edit_comment.save ? 'Success' : 'Fail'
+      else
+        @code = 'NoPermission'
       end
     end
     @code ||= 'Fail'
@@ -67,16 +60,14 @@ class Index::Workspace::EditCommentsController < IndexController
   end
 
   def destroy
-    unless @code
-      # 允许删除评论的条件
-      if @edit_comment.user == @user || # 1: 评论者
-        @user.can_edit?(:delete_comment, @resource) # 2: 具有删除评论权限的用户
+    # 允许删除评论的条件
+    @code = if @edit_comment.user == @user || # 1: 评论者
+               @user.can_edit?(:delete_comment, @resource) # 2: 具有删除评论权限的用户
 
-        @code = @edit_comment.destroy ? 'Success' : 'Fail'
-      else
-        @code = 'NoPermission'
-      end
-    end
+              @edit_comment.destroy ? 'Success' : 'Fail'
+            else
+              'NoPermission'
+            end
     render json: { code: @code }
   end
 
@@ -84,8 +75,8 @@ class Index::Workspace::EditCommentsController < IndexController
 
   # Use callbacks to share common setup or constraints between actions.
   def set_edit_comment
-    @edit_comment = @resource.edit_comments.find_by_id params[:id] if @resource
-    @code ||= 'ResourceNotExist' unless @edit_comment
+    edit_comment_cache if @resource
+    render(json: { code: 'ResourceNotExist' }) && return unless @edit_comment
   end
 
   # Never trust parameters from the scary internet, only allow the white list through.
@@ -93,16 +84,17 @@ class Index::Workspace::EditCommentsController < IndexController
     params.require(:edit_comment).permit(:hash_key, :content)
   end
 
-  def set_resource
-    if @user
-      resource_type = params[:resource_type]
-      resource_id = params[:resource_id]
-      @resource = case resource_type
-                  when 'articles'
-                    Index::Workspace::Article.shown.find_by_id resource_id
-                  end
-      @code ||= 'ResourceNotExist' unless @resource
-      @code ||= 'NoPermission' && @resource = nil if @resource && !@user.can_edit?(:comment, @resource)
-    end
+  def set_resource # 之后可能会添加其它的编辑评论对象，此方法方便扩展
+    resource_type = params[:resource_type]
+    resource_id = params[:resource_id]
+    @resource = case resource_type
+                when 'articles'
+                  edit_article_cache resource_id
+                end
+    render(json: { code: 'ResourceNotExist' }) && return unless @resource
+  end
+
+  def check_permission
+    render(json: { code: 'NoPermission' }) unless @user.can_edit?(:comment, @resource)
   end
 end
