@@ -1,4 +1,4 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 require 'rails/all'
 
@@ -8,6 +8,9 @@ Bundler.require(*Rails.groups)
 
 module Server
   class Application < Rails::Application
+    # Initialize configuration defaults for originally generated Rails version.
+    config.load_defaults 5.1
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
@@ -16,16 +19,17 @@ module Server
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
     config.active_record.default_timezone = :local
-    config.time_zone = 'Beijing'  
+    config.time_zone = 'Beijing'
 
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
 
-    # Do not swallow errors in after_commit/after_rollback callbacks.
-    config.active_record.raise_in_transactional_callbacks = true
+    #使用 schema_search_path 或者其他 PostgreSQL 扩展，可以控制如何转储数据库模式
+    config.active_record.dump_schemas = :all
 
-    config.cache_store = :redis_store,  'redis://localhost:6379/0/cache'
+    config.cache_store = :dalli_store
+    # config.cache_store = :redis_store,  'redis://localhost:6379/0/cache'
     # , {
     #   host: "localhost",
     #   port: 6379,
@@ -45,10 +49,10 @@ module Server
     # Avoid CORS issues when API is called from the frontend app.
     # Handle Cross-Origin Resource Sharing (CORS) in order to accept cross-origin AJAX requests.
     # Read more: https://github.com/cyu/rack-cors
-    config.middleware.insert_before 0, 'Rack::Cors' do
+    config.middleware.insert_before 0, Rack::Cors do
       allow do
         origins '*'
-        resource '*', headers: :any, methods: [:get, :post, :put, :patch, :delete, :head, :options]
+        resource '*', :headers => :any, :methods => [:get, :post, :put, :patch, :delete, :options]
       end
     end
   end
